requirements.txt
# CrewAI Project Requirements

# ============================================================================
# Core Dependencies
# ============================================================================
crewai>=0.86.0
crewai-tools>=0.17.0
agentops>=0.3.14

# ============================================================================
# LLM Providers
# ============================================================================
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# ============================================================================
# RAG Dependencies
# ============================================================================
# Vector store and embeddings
faiss-cpu>=1.7.4  # For CPU-only systems (use faiss-gpu for GPU)
numpy>=1.24.0,<2.0.0  # FAISS requires numpy 1.x

# Document processing
pypdf>=4.0.0  # PDF support (optional)
python-docx>=1.1.0  # DOCX support (optional)

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0

# ============================================================================
# Optional but Recommended
# ============================================================================
rich>=13.0.0  # Better terminal output

# ============================================================================
# Development Dependencies (uncomment for development)
# ============================================================================
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
